requirements.txt
# SQL inject parameters
jinja2==3.1.6

# Google connectivity
google_oauth2_tool
google-cloud-bigquery
functions-framework==3.*
